% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust-perturb.R
\name{clust.perturb}
\alias{clust.perturb}
\title{Perturb clusters}
\usage{
clust.perturb(
  network,
  clustering.algorithm,
  noise = 0.1,
  iters = 3,
  edge.list.format = NULL,
  cluster.format = NULL
)
}
\arguments{
\item{clustering.algorithm}{either a character string specifying one of four clustering
algorithms ("mcl", "walktrap", "hierarchical", "k-med") or a function responsible for
clustering}

\item{noise}{vector with values between 0 and 1 specifying the amount(s) of noise to 
add to the network.}

\item{iters}{positive integer specifying number of iterations.}

\item{edge.list.format}{optional function that transforms edge list into format required
by clustering.algorithm.}

\item{cluster.format}{optional function that transforms output returned by 
clustering.algorithm into a data frame with three or four columns, where columns one 
and two are an edge list, column three is an id specifying the cluster assignment, and}

\item{edge.list}{data frame with two or three columns, with first two columns providing 
the list of edges and an optional third column giving edge weight.}
}
\value{
data frame containing all clusters and their density score.
}
\description{
Network to be clustered is formated as a data frame edge list (edge.list). Users can 
pass a custom function edge.list.format to handle different formats required by
clustering.algorithm. edge.list.format must transform a data frame edge list (two 
columns unweighted, or three columsn weighted) into the required format.
}
\details{
Assumes clustering.algorithm returns a list of clustered nodes. If clustering.algorithm
returns a different format, users can pass a custom function cluster.format that
transforms the output of clustering.algorithm into a list of clustered nodes.
}
\examples{
first example

second example
}
